Couch user profiles: aliases.view -> users.design (with `list` function instead of `reduce` one)
